stories as jsx
