+++ package.json
@@ -82,12 +82,12 @@
   "devDependencies": {
     "@po.et/stylelint-rules": "1.1.0",
     "@po.et/tslint-rules": "1.2.0",
-    "@storybook/addon-actions": "^3.4.5",
-    "@storybook/addon-info": "^3.4.5",
-    "@storybook/addon-knobs": "^3.4.5",
-    "@storybook/addon-links": "^3.4.5",
-    "@storybook/addons": "^3.4.5",
-    "@storybook/react": "^3.4.5",
+    "@storybook/addon-actions": "^3.4.6",
+    "@storybook/addon-info": "^3.4.6",
+    "@storybook/addon-knobs": "^3.4.6",
+    "@storybook/addon-links": "^3.4.6",
+    "@storybook/addons": "^3.4.6",
+    "@storybook/react": "^3.4.6",
     "@types/babel-core": "6.25.3",
     "@types/classnames": "2.2.3",
     "@types/copy-webpack-plugin": "4.4.1",
@@ -105,6 +105,11 @@
     "@types/redux-devtools": "3.0.44",
     "@types/redux-persist": "4.3.1",
     "@types/redux-saga": "0.10.5",
+    "@types/storybook__addon-actions": "3.0.3",
+    "@types/storybook__addon-info": "3.2.3",
+    "@types/storybook__addon-knobs": "3.3.1",
+    "@types/storybook__addon-links": "3.3.0",
+    "@types/storybook__react": "3.0.7",
     "@types/webpack": "4.1.4",
     "@types/webpack-dev-middleware": "2.0.1",
     "@types/webpack-dev-server": "2.9.4",
@@ -112,7 +117,6 @@
     "awesome-typescript-loader": "3.0.0-beta.17",
     "babel-core": "6.26.3",
     "babel-loader": "6.4.1",
-    "babel-runtime": "^6.26.0",
     "copy-webpack-plugin": "4.5.1",
     "css-loader": "0.28.11",
     "express": "4.16.3",
