+++ src/components/atoms/Button/Button.stories.tsx
@@ -1,14 +1,13 @@
-import React from 'react'
+import * as React from 'react'
 
+import { text, boolean, select } from '@storybook/addon-knobs/react'
 import { storiesOf } from '@storybook/react'
-import { Button } from './Button'
+// tslint:disable-next-line
 import { wInfo } from '../../../stories/index.stories'
-import { text, boolean } from '@storybook/addon-knobs/react'
+import { Button } from './Button'
 
-
-
-storiesOf('Components/Button', module).addWithJSX(
-  'Basic Button',
+storiesOf('Components/Buttons', module).addWithJSX(
+  'Button',
   wInfo(`
 
   ### Notes
@@ -21,17 +20,16 @@ storiesOf('Components/Button', module).addWithJSX(
     readonly className?: string
     readonly text?: string
     readonly disabled?: boolean
-    readonly onClick?: (event: Event) => void
+    readonly onClick?: any
     readonly type?: 'primary' | 'danger'
     />
-  ~~~`
-)(() => (
+  ~~~`)(() => (
     <Button
+      className={text('className', '')}
       text={text('text', 'Test Button')}
       disabled={boolean('disabled', false)}
-      type={text('type','primary')}
+      type={select('type', { primary: 'primary', danger: 'danger' }, 'primary')}
       onClick={() => alert('clicked')}
-      className={text('className')}
-     />
+    />
   ))
 )
